requirements.txt
fastapi
uvicorn
pydantic
requests
python-dotenv
numpy
scipy
PyJWT